requirements.txt
alembic==1.11.1
aniso8601==9.0.1
blinker==1.6.2
click==8.1.3
Flask==2.3.2
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.19.0
mysqlclient==2.1.1
packaging==23.1
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.16
typing_extensions==4.6.3
Werkzeug==2.3.6
